[add] Enemyの攻撃処理をEnemyAttackクラスに分離した。
