+++ Assets/Scripts/Enemy/ProtoEnemyAttack.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e8383f6b4876ec3448acc7e2b828c802
+guid: 9a2fbb7a6b908114bb2c46752ae3a4b3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
